Remove '.tmux.conf', some aliases, and Syntastic; add 'lesspipe'
I had tried to use 'tmux' on my servers and had remained disappointed.
It looks pretty ugly, despite the configuration. Vim doesn't have a
background color within tmux. It's just rendered as a black screen with
some text. I don't like it. Since I don't know how to tweak it more
properly, I've decided to stop using it.

The second issue relates to Syntastic -- a syntax checker for Vim. I
wanted to use it for Python but encountered some problems. The main one
is that the python executable crashes out with "abnormal termination"
if there are any Cyrillic characters in the code. It's very bad and
embarassing, I think, to have such issues in 2018, but it does. I don't
know how to fix it, and I just won't use Syntastic anymore in favor of
already properly pre-configured and well-working IDEs such as PyCharm
by JetBrains.

Lastly, I think it was silly to write aliases for short commands to
make them even shorter. It's better to adhere to the common conventions
and program names, and not over-inventing things like that.